+++ src/Controller/Profile/HomeController.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace App\Controller\Profile;
+
+use Doctrine\ORM\EntityManagerInterface;
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Routing\Attribute\Route;
+
+final class HomeController extends AbstractController
+{
+    #[Route('/profile', name: 'app_profile')]
+    public function index(): Response
+    {
+        return $this->render('profile/index.html.twig');
+    }
+}
