+++ src/Controller/Profile/PasswordController.php
@@ -1,0 +1,39 @@
+<?php
+
+namespace App\Controller\Profile;
+
+use App\Form\PasswordUserType;
+use Doctrine\ORM\EntityManagerInterface;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\PasswordHasher\Hasher\UserPasswordHasherInterface;
+
+class PasswordController extends AbstractController
+{
+    private $entityManager;
+    public function __construct(EntityManagerInterface $entityManager)
+    {
+        $this->entityManager = $entityManager;
+    }
+    
+    #[Route('/profile/change-password', name: 'app_profile_change_password')]
+    public function index(Request $request, UserPasswordHasherInterface $passwordHasher): Response
+    {
+        $user = $this->getUser();
+        $form = $this->createForm(PasswordUserType::class, $user, ['passwordHasher' => $passwordHasher]);
+        $form->handleRequest($request);
+        if ($form->isSubmitted() && $form->isValid()) {
+            $this->entityManager->flush();
+            $this->addFlash(
+                'success',
+                'Your password have been updated'
+            );
+        }
+
+        return $this->render('profile/password/index.html.twig', [
+            'form' => $form
+        ]);
+    }
+}
